In double, allow to choose who serves the first.
In double, players of the same team take turn to serve. The new option allows to choose who among the watch carrier or his teammate will be the first to serve.
To accommodate this new option, create a new class to store the configuration of the match before its creation.