Add endless mode.
It is now possible to start a match in endless mode. In this mode, an infinite number (watch memory is the limit) of sets can be played and the user can end a match anytime. When the match is ended, a winner is determined, according to the following winning rules:
-if the number of sets won by the teams is different, the team with the highest number of sets won wins the match (the current set is not considered)
-if both teams have won the same number of sets, the team with the highest total score wins the match (the current set is not considered)
-if both teams have the same number of sets won and the same total score, the match is a draw

Moreover, in standard mode, the user can now end prematurely a match using the match menu. In this case, the winner will be determined using the same winning rules as for endless matches.

Technically, the following has been done:
-when creating a match, the maximum number of sets is stored (this number is null in endless mode)
-a list is used to store the match sets because the number of sets is unknown at the beginning of the match
-a specific boolean is set to true when the match has ended because it is no longer possible to rely on the winner flag (because a match can be a draw)
-the activity session is now updated after each rally (instead of at the end of a set) so it is up to date when the user ends the match in the middle of a set
-the application handle draws