Use a perspective for the court even when it is not necessary.
Previously, the court was displayed using all the available space. This was resulting in a rectangular court when the time was displayed because the top and bottom margins were the same. Now, a perspective is always used to improve aesthetics.

The shape and the dimension of the court have also been updated.