Update the set scores only once.
This resolves an issue with the Garmin activity in endless mode: if the match was ended with a pending blank set (a set without any rally), the scores of the previous set were used for the blank set, seemingly creating two identical laps in the activity. Now, the set scores are updated when a set ends, or when the match is manually ended. It's better to set lap and session data only once.